Fix de Base de Datos Fix Login Google+ Fix Login Facebook Cambio de Iconos

+++ Android/AppRunning/mobile/src/main/java/studios/thinkup/com/apprunning/adapter/DetalleCarreraPagerAdapter.java
@@ -12,10 +12,10 @@ import studios.thinkup.com.apprunning.model.entity.UsuarioCarrera;
  * Created by fcostazini on 21/05/2015.
  */
 public class DetalleCarreraPagerAdapter extends FragmentPagerAdapter {
-    private long carrera;
+    private int carrera;
 
 
-    public DetalleCarreraPagerAdapter(FragmentManager fm, long idCarrera) {
+    public DetalleCarreraPagerAdapter(FragmentManager fm, int idCarrera) {
         super(fm);
         this.carrera = idCarrera;
     }
